refactor: PostCreate를 위한 훅 생성
